+++ DynamicQuerying/DynamicQuerying.sln
@@ -6,6 +6,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamicQuerying.Tests", "Dy
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamicQuerying.Sample", "DynamicQuerying.Sample\DynamicQuerying.Sample.csproj", "{0044EE1E-DE36-4915-9A80-2DDCC7274DA8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DynamicQuerying.App", "DynamicQuerying.App\DynamicQuerying.App.csproj", "{3E20EBE4-79AF-42DE-90C9-A56D7E285F81}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -24,5 +26,9 @@ Global
 		{0044EE1E-DE36-4915-9A80-2DDCC7274DA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0044EE1E-DE36-4915-9A80-2DDCC7274DA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0044EE1E-DE36-4915-9A80-2DDCC7274DA8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3E20EBE4-79AF-42DE-90C9-A56D7E285F81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3E20EBE4-79AF-42DE-90C9-A56D7E285F81}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3E20EBE4-79AF-42DE-90C9-A56D7E285F81}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3E20EBE4-79AF-42DE-90C9-A56D7E285F81}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
